+++ packages/ragbits-chat/src/ragbits/chat/interface/_interface.py
@@ -9,12 +9,22 @@ from abc import ABC, abstractmethod
 from collections.abc import AsyncGenerator, Callable
 from typing import Any, Literal
 
+from ragbits.chat.interface.ui_customization import UICustomization
 from ragbits.core.prompt.base import ChatFormat
 from ragbits.core.utils import get_secret_key
 
 from ..persistence import HistoryPersistenceStrategy
 from .forms import FeedbackConfig
-from .types import ChatContext, ChatResponse, ChatResponseType, Reference, StateUpdate
+from .types import (
+    ChatContext,
+    ChatResponse,
+    ChatResponseType,
+    LiveUpdate,
+    LiveUpdateContent,
+    LiveUpdateType,
+    Reference,
+    StateUpdate,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -85,6 +95,7 @@ class ChatInterface(ABC):
     """
 
     feedback_config: FeedbackConfig = FeedbackConfig()
+    ui_customization: UICustomization | None = None
     history_persistence: HistoryPersistenceStrategy | None = None
 
     def __init_subclass__(cls, **kwargs: dict) -> None:
@@ -118,6 +129,23 @@ class ChatInterface(ABC):
             type=ChatResponseType.STATE_UPDATE,
             content=StateUpdate(state=state, signature=signature),
         )
+
+    @staticmethod
+    def create_live_update(
+        update_id: str, type: LiveUpdateType, label: str, description: str | None = None
+    ) -> ChatResponse:
+        """Helper method to create a live update response."""
+        return ChatResponse(
+            type=ChatResponseType.LIVE_UPDATE,
+            content=LiveUpdate(
+                update_id=update_id, type=type, content=LiveUpdateContent(label=label, description=description)
+            ),
+        )
+
+    @staticmethod
+    def create_followup_messages(messages: list[str]) -> ChatResponse:
+        """Helper method to create a live update response."""
+        return ChatResponse(type=ChatResponseType.FOLLOWUP_MESSAGES, content=messages)
 
     @staticmethod
     def _sign_state(state: dict[str, Any]) -> str: